delete idle instances

+++ ci/gcloud-init.sh
@@ -113,10 +113,12 @@ for manifest in $(ls ${script_dir}/../userdata/Manifest/*-gamma.json); do
     if [ -n "${zoneUrl}" ] && [[ "${zoneUrl}" != "null" ]]; then
       zone=${zoneUrl##*/}
       latestResolvedTaskTimeInUtc=$(curl -s "https://queue.taskcluster.net/v1/task/$(_jq '.latestTask.taskId')/status" | jq --arg runId $(_jq '.latestTask.runId') -r '.status.runs[] | select(.runId == ($runId | tonumber)) | .resolved')
-      latestResolvedTaskTime=$(date --date "${latestResolvedTaskTimeInUtc}" +%s)
-      echo "$(tput dim)[${script_name} $(date --utc +"%F %T.%3NZ")]$(tput sgr0) ${zone}/$(_jq '.workerId') last resolved task: $(tput bold)${latestResolvedTaskTimeInUtc}$(tput sgr0) ($(( ($(date +%s) - $latestResolvedTaskTime) / 60)) minutes ago)"
-      if [ ${latestResolvedTaskTime} -lt ${idleThreshold} ] && gcloud compute instances delete $(_jq '.workerId') --zone ${zone} --delete-disks all --quiet; then
-        echo "$(tput dim)[${script_name} $(date --utc +"%F %T.%3NZ")]$(tput sgr0) deleted: $(tput bold)${zone}/$(_jq '.workerId')$(tput sgr0)"
+      if [ -n "${latestResolvedTaskTimeInUtc}" ] && [[ "${latestResolvedTaskTimeInUtc}" != "null" ]]; then
+        latestResolvedTaskTime=$(date --date "${latestResolvedTaskTimeInUtc}" +%s)
+        echo "$(tput dim)[${script_name} $(date --utc +"%F %T.%3NZ")]$(tput sgr0) ${zone}/$(_jq '.workerId') last resolved task: $(tput bold)${latestResolvedTaskTimeInUtc}$(tput sgr0) ($(( ($(date +%s) - $latestResolvedTaskTime) / 60)) minutes ago)"
+        if [ ${latestResolvedTaskTime} -lt ${idleThreshold} ] && gcloud compute instances delete $(_jq '.workerId') --zone ${zone} --delete-disks all --quiet; then
+          echo "$(tput dim)[${script_name} $(date --utc +"%F %T.%3NZ")]$(tput sgr0) deleted: $(tput bold)${zone}/$(_jq '.workerId')$(tput sgr0)"
+        fi
       fi
     fi
   done
